+++ test/libsolidity/syntaxTests/types/mapping/mapping_data_location_function_param_external.sol
@@ -2,5 +2,5 @@ contract c {
     function f1(mapping(uint => uint) calldata) pure external returns (mapping(uint => uint) memory) {}
 }
 // ----
-// TypeError: (29-59): Mapping types for parameters or return variables can only be used in internal or library functions.
-// TypeError: (84-112): Mapping types for parameters or return variables can only be used in internal or library functions.
+// TypeError: (29-59): Mapping types can only have a data location of "storage" and thus only be parameters or return variables for internal or library functions.
+// TypeError: (84-112): Mapping types can only have a data location of "storage" and thus only be parameters or return variables for internal or library functions.