+++ test/libsolidity/syntaxTests/types/mapping/mapping_data_location_function_param_internal.sol
@@ -1,4 +1,6 @@
 contract c {
-    function f4(mapping(uint => uint) memory) pure internal {}
+    function f4(mapping(uint => uint) storage) pure internal {}
+    function f5(mapping(uint => uint) memory) pure internal {}
 }
 // ----
+// TypeError: (93-121): Mapping types can only have a data location of "storage".